Add functionality to check and notify about F1 races
The code updates include the addition of a ticker that checks for F1 race notifications every minute. Two functions were added: `checkF1Notification` which checks if a race is about to start and sends a message if so, and `getNextMessageWithCache` which retrieves race data from an API, adding a caching layer for performance. Minor updates to string checks in certain functions were also done.